fix: correcao tipo do id Role

+++ Core/IBGEExplorer.Account/Entities/Role.cs
@@ -2,7 +2,7 @@
 
 namespace IBGEExplorer.Account.Entities;
 
-public class Role : Entity<int>
+public class Role : Entity<short>
 {
     public string? Name { get; set; }
